requirements.txt
streamlit
python-dotenv
google-generativeai
extra-streamlit-components